Corregido error ortográfico

+++ CustomFaster_MP_TPU/roi_heads.py
@@ -35,7 +35,7 @@ def fastrcnn_loss(class_logits, box_regression, labels, regression_targets):
     regression_targets = torch.cat(regression_targets, dim=0)  
 
     class_weights = torch.tensor([0.5, 2]).to(device)
-    classification_loss = F.cross_entropy(class_logits, labels.long(), weight=class_weights)
+    classification_loss = F.cross_entropy(class_logits, labels.float(), weight=class_weights)
 
     # get indices that correspond to the regression targets for
     # the corresponding ground truth labels, to be used with
